Add methods for player input and insert to board
get_choice prompts the player to input a number from 1 to 9, and loops
infinitely if the incorrect response type is given.

insert_symbol places one of the player's symbols (X or O) on the board,
and is passed an integer value corresponding to the index value on the
board where the symbol is to be placed.