Create gameplay loop and slightly alter board ary
Put the main turn-taking logic into a method that loops; here is where
additional logic will be added later, such as switching who's turn it
is.

The board array has been increased from size 9 to 10, not because the
board needs to be bigger but for convenience sake. Index 0 will be
ignored on purpose; the player's chosen input will directly translate
to a space on the board, removing the need to subtract 1 from their
input, as well as making it less confusing to make their choice.